Distribution to Dev.to/requirements.txt
openai-agents>=0.6.3
requests>=2.32.3
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
Flask>=3.0.3
google-genai>=1.3.0
Pillow>=10.4.0
